backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
anthropic>=0.18.0
python-dotenv==1.0.0
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2
pydantic>=2.0.0
python-dotenv==1.0.0
